requirements.txt
streamlit
opencv-python
mediapipe
numpy
imutils
matplotlib
deepface
protobuf==3.20.3
Pillow
tf-keras
